requirements.txt
fastapi
uvicorn
mysql-connector-python
groq
requests
streamlit
geocoder
pydantic